+++ crates/plugin_partial_bundling/src/lib.rs
@@ -24,6 +24,10 @@ impl Plugin for FarmPluginPartialBundling {
     "FarmPluginPartialBundling"
   }
 
+  fn priority(&self) -> i32 {
+    99
+  }
+
   fn analyze_module_graph(
     &self,
     module_graph: &mut ModuleGraph,
@@ -54,6 +58,10 @@ impl Plugin for FarmPluginPartialBundling {
 
     for module_id in modules {
       let module = module_graph.module(module_id).unwrap();
+      // Skip the external modules
+      if module.external {
+        continue;
+      }
 
       if module.resource_pot.is_some() {
         panic!(