+++ crates/plugin_react/src/lib.rs
@@ -3,11 +3,15 @@
 use farmfe_core::{
   config::Config,
   plugin::{Plugin, PluginAnalyzeDepsHookResultEntry, ResolveKind},
-  swc_common::{comments::NoopComments, Mark, GLOBALS},
+  swc_common::{comments::NoopComments, errors::HANDLER, Mark, GLOBALS},
 };
 
 use farmfe_toolkit::{
-  swc_ecma_transforms::react::{react, Options, RefreshOptions},
+  script::swc_try_with::try_with,
+  swc_ecma_transforms::{
+    helpers::inject_helpers,
+    react::{react, Options, RefreshOptions},
+  },
   swc_ecma_visit::VisitMutWith,
 };
 use react_refresh::inject_react_refresh;
@@ -128,31 +132,37 @@ impl Plugin for FarmPluginReact {
       param.module_type,
       farmfe_core::module::ModuleType::Jsx | farmfe_core::module::ModuleType::Tsx
     ) {
-      GLOBALS.set(&context.meta.script.globals, || {
-        let top_level_mark = Mark::from_u32(param.meta.as_script().top_level_mark);
-        let ast = &mut param.meta.as_script_mut().ast;
-        let is_dev = matches!(context.config.mode, farmfe_core::config::Mode::Development);
+      try_with(
+        context.meta.script.cm.clone(),
+        &context.meta.script.globals,
+        || {
+          let top_level_mark = Mark::from_u32(param.meta.as_script().top_level_mark);
+          let unresolved_mark = Mark::from_u32(param.meta.as_script().unresolved_mark);
+          let ast = &mut param.meta.as_script_mut().ast;
+          let is_dev = matches!(context.config.mode, farmfe_core::config::Mode::Development);
 
-        ast.visit_mut_with(&mut react(
-          context.meta.script.cm.clone(),
-          Some(NoopComments), // TODO parse comments
-          Options {
-            refresh: if is_dev {
-              Some(RefreshOptions::default())
-            } else {
-              None
+          ast.visit_mut_with(&mut react(
+            context.meta.script.cm.clone(),
+            Some(NoopComments), // TODO parse comments
+            Options {
+              refresh: if is_dev {
+                Some(RefreshOptions::default())
+              } else {
+                None
+              },
+              development: Some(is_dev),
+              // runtime: Some(Runtime::Automatic),
+              ..Default::default()
             },
-            development: Some(is_dev),
-            // runtime: Some(Runtime::Automatic),
-            ..Default::default()
-          },
-          top_level_mark,
-        ));
+            top_level_mark,
+          ));
+          ast.visit_mut_with(&mut inject_helpers(unresolved_mark));
 
-        if is_dev {
-          inject_react_refresh(ast);
-        }
-      });
+          if is_dev {
+            inject_react_refresh(ast);
+          }
+        },
+      )?;
 
       return Ok(Some(()));
     }