+++ crates/plugin_static_assets/src/lib.rs
@@ -3,13 +3,16 @@
 use std::path::Path;
 
 use base64::engine::{general_purpose, Engine};
-use farmfe_core::{config::Config, module::ModuleType, plugin::Plugin, resource::{Resource, ResourceType}};
+use farmfe_core::{
+  config::Config,
+  module::ModuleType,
+  plugin::Plugin,
+  resource::{Resource, ResourceType},
+};
 use farmfe_toolkit::{
   fs::{read_file_raw, read_file_utf8, transform_output_filename},
   lazy_static::lazy_static,
 };
-
-const VIRTUAL_ASSET_PREFIX: &str = "virtual:FARM_ASSETS:";
 
 // Default supported static assets: png, jpg, jpeg, gif, svg, webp, mp4, webm, wav, mp3, wma, m4a, aac, ico, ttf, woff, woff2
 lazy_static! {
@@ -33,7 +36,7 @@ impl Plugin for FarmPluginStaticAssets {
   }
   /// Make sure this plugin is executed last
   fn priority(&self) -> i32 {
-    0
+    99
   }
 
   fn resolve(
@@ -81,7 +84,7 @@ impl Plugin for FarmPluginStaticAssets {
     context: &std::sync::Arc<farmfe_core::context::CompilationContext>,
   ) -> farmfe_core::error::Result<Option<farmfe_core::plugin::PluginTransformHookResult>> {
     if matches!(param.module_type, ModuleType::Asset) {
-      if param.query.contains_key("inline") {
+      if param.query.iter().find(|(k, _)| k == "inline").is_some() {
         let file_raw = read_file_raw(param.resolved_path)?;
         let file_base64 = general_purpose::STANDARD.encode(&file_raw);
         let path = Path::new(param.resolved_path);
@@ -97,7 +100,7 @@ impl Plugin for FarmPluginStaticAssets {
           module_type: Some(ModuleType::Js),
           source_map: None,
         }));
-      } else if param.query.contains_key("raw") {
+      } else if param.query.iter().find(|(k, _)| k == "raw").is_some() {
         let file_utf8 = read_file_utf8(param.resolved_path)?;
         let content = format!("export default \"{}\"", file_utf8);
 
@@ -126,14 +129,17 @@ impl Plugin for FarmPluginStaticAssets {
         let content = format!("export default \"/{}\"", resource_name);
 
         let mut resources_map = context.resources_map.lock();
-        resources_map.insert(resource_name.clone(), Resource {
+        resources_map.insert(
+          resource_name.clone(),
+          Resource {
             name: resource_name,
             bytes,
             emitted: false,
             resource_type: ResourceType::Asset(ext.to_string()),
             resource_pot: "STATIC_ASSETS".into(),
             preserve_name: false,
-        });
+          },
+        );
 
         return Ok(Some(farmfe_core::plugin::PluginTransformHookResult {
           content,