+++ examples/script-entry/index.ts
@@ -1,9 +1,12 @@
+import module from 'module';
+
 import { UserConfig } from '@farmfe/core';
 
 export default {
   compilation: {
     input: {
-      main: './main.tsx',
+      main: './main.tsx'
     },
-  },
+    external: module.builtinModules
+  }
 } as UserConfig;