+++ packages/core/binding/index.d.ts
@@ -1,11 +1,11 @@
 export type ModuleType =
-  | 'Ts'
-  | 'Js'
-  | 'Jsx'
-  | 'Tsx'
-  | 'Css'
-  | 'Html'
-  | 'Asset'
+  | 'ts'
+  | 'js'
+  | 'jsx'
+  | 'tsx'
+  | 'css'
+  | 'html'
+  | 'asset'
   | string;
 
 export type ResolveKind =
@@ -26,12 +26,13 @@ export default Compiler;
 
 /// Parameter of the resolve hook
 export interface PluginResolveHookParam {
-  /// the specifier would like to resolve, for example, './index'
-  specifier: String;
-  /// the start location to resolve `specifier`, being [None] if resolving a entry or resolving a hmr update.
-  importer: String | null;
+  /// the start location to resolve `source`, being [None] if resolving a entry or resolving a hmr update.
+  importer: { relativePath: string; queryString: string | null } | null;
   /// for example, [ResolveKind::Import] for static import (`import a from './a'`)
   kind: ResolveKind;
+  /// resolvedPath. for example in index.ts (import App from "./App.vue")
+  /// source should be "path.resolve(process.cwd(),'./App.vue')"
+  source: String;
 }
 
 export interface PluginResolveHookResult {
@@ -44,12 +45,15 @@ export interface PluginResolveHookResult {
   /// the query parsed from specifier, for example, query should be `{ inline: true }` if specifier is `./a.png?inline`
   /// if you custom plugins, your plugin should be responsible for parsing query
   /// if you just want a normal query parsing like the example above, [crate::utils::parse_query] is for you
-  query: Record<string, string> | null;
+  query: [string, string][] | null;
+  /// meta data of the module, will be passed to [PluginLoadHookParam] and [PluginTransformHookParam]
+  meta: Record<string, string> | null;
 }
 
 export interface PluginLoadHookParam {
   resolvedPath: string;
-  query: Record<string, string>;
+  query: [string, string][];
+  meta: Record<string, string> | null;
 }
 
 export interface PluginLoadHookResult {
@@ -66,7 +70,8 @@ export interface PluginTransformHookParam {
   /// module type after load
   moduleType: ModuleType;
   resolvedPath: string;
-  query: Record<string, string>;
+  query: [string, string][];
+  meta: Record<string, string> | null;
 }
 
 export interface PluginTransformHookResult {
@@ -85,13 +90,14 @@ export interface Config {
       filename?: string;
       path?: string;
       publicPath?: string;
+      assetsFilename?: string;
     };
     resolve?: {
       extensions?: string[];
       alias?: Record<string, string>;
       mainFields?: string[];
       conditions?: string[];
-      symlinks: boolean;
+      symlinks?: boolean;
     };
     define?: Record<string, string>;
     external?: string[];
@@ -146,6 +152,7 @@ export interface Config {
         };
       };
     };
+    sourcemap?: boolean | 'all';
     partialBundling?: {
       moduleBuckets?: {
         name: string;