+++ packages/runtime-plugin-hmr/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @farmfe/runtime-plugin-hmr
 
+## 3.0.5
+
+### Patch Changes
+
+- limit the watched files to optimize cold start speed and fix lazy compilation issue"
+
+## 3.0.4
+
+### Patch Changes
+
+- Fix swc helper inject issue and optimize CLI
+
+## 3.0.3
+
+### Patch Changes
+
+- 98d662f: feat: reconnect hmr server
+
 ## 3.0.2
 
 ### Patch Changes