+++ packages/runtime/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @farmfe/runtime
 
+## 0.3.4
+
+### Patch Changes
+
+- Auto external node native module when reading farm.config.ts
+
+## 0.3.3
+
+### Patch Changes
+
+- limit the watched files to optimize cold start speed and fix lazy compilation issue"
+
 ## 0.3.2
 
 ### Patch Changes